// Angular-Hospital-Management-System/src/app/shared/model/user.ts
export interface User {
    userId: number; // Optional, usually generated by backend
    firstName: string;
    middleName?: string;
    lastName: string;
    email: string;
    password: string;
    phone: string;
    role: 'USER'; // Match with your UserRole Enum
    gender?: 'MALE' | 'FEMALE' | 'OTHER'; // Match your Gender Enum
    photoURL?: string;
    dateOfBirth: string; // ISO format: 'yyyy-MM-dd'
    appointment?: any[]; // Optionally typed with Appointment[] if interface is ready
    prescriptions?: any[]; // Optionally typed with Prescription[] if interface is ready

  }
  